+++ visualization/app/codeCharta/state/effects/nodeContextMenu/nodeContextMenu.e2e.ts
@@ -1,7 +1,7 @@
 import { clearIndexedDB, goto } from "../../../../puppeteer.helper"
 import { NodeContextMenuPageObject } from "./nodeContextMenu.po"
-import { SearchPanelModeSelectorPageObject } from "../../../ui/searchPanel/searchPanelModeSelector/searchPanelModeSelector.po"
-import { MapTreeViewLevelPageObject } from "../../../ui/searchPanel/mapTreeView/mapTreeView.level.po"
+import { SearchPanelModeSelectorPageObject } from "../../../ui/ribbonBar/searchPanel/searchPanelModeSelector/searchPanelModeSelector.po"
+import { MapTreeViewLevelPageObject } from "../../../ui/ribbonBar/searchPanel/mapTreeView/mapTreeView.level.po"
 import { CodeMapPageObject } from "../../../ui/codeMap/codeMap.po"
 
 describe("NodeContextMenu", () => {