+++ visualization/app/codeCharta/ui/ribbonBar/edgeMetricChooser/edgeMetricChooser.module.ts
@@ -1,11 +1,12 @@
 import { CommonModule } from "@angular/common"
 import { NgModule } from "@angular/core"
 import { MetricChooserModule } from "../../metricChooser/metricChooser.module"
-import { MetricChooserTypeModule } from "../../metricChooser/metricChooserType/metricChooserType.module"
 import { EdgeMetricChooserComponent } from "./edgeMetricChooser.component"
+import { MetricChooserTypeModule } from "../../metricChooser/metricChooserType/metricChooserType.module"
+import { RoundedBoxModule } from "../roundedBox/roundedBox.module"
 
 @NgModule({
-    imports: [CommonModule, MetricChooserModule, MetricChooserTypeModule],
+    imports: [CommonModule, MetricChooserModule, MetricChooserTypeModule, RoundedBoxModule],
     declarations: [EdgeMetricChooserComponent],
     exports: [EdgeMetricChooserComponent]
 })