+++ visualization/app/codeCharta/ui/ribbonBar/heightMetricChooser/heightMetricChooser.module.ts
@@ -1,11 +1,10 @@
 import { CommonModule } from "@angular/common"
 import { NgModule } from "@angular/core"
 import { MetricChooserModule } from "../../metricChooser/metricChooser.module"
-import { MetricChooserValueModule } from "../../metricChooser/metricChooserValue/metricChooserValue.module"
 import { HeightMetricChooserComponent } from "./heightMetricChooser.component"
 
 @NgModule({
-    imports: [CommonModule, MetricChooserModule, MetricChooserValueModule],
+    imports: [CommonModule, MetricChooserModule],
     declarations: [HeightMetricChooserComponent],
     exports: [HeightMetricChooserComponent]
 })