+++ visualization/app/codeCharta/ui/ribbonBar/ribbonBar.module.ts
@@ -7,7 +7,6 @@ import { EdgeMetricChooserModule } from "./edgeMetricChooser/edgeMetricChooser.m
 import { ShowScenariosButtonModule } from "./showScenariosButton/showScenariosButton.module"
 import { EdgeSettingsPanelModule } from "./edgeSettingsPanel/edgeSettingsPanel.module"
 import { ArtificialIntelligenceModule } from "./artificialIntelligence/artificialIntelligence.module"
-import { SearchPanelModule } from "../searchPanel/searchPanel.module"
 import { AreaSettingsPanelModule } from "./areaSettingsPanel/areaSettingsPanel.module"
 import { CustomConfigsModule } from "../customConfigs/customConfigs.module"
 import { HeightSettingsPanelModule } from "./heightSettingsPanel/heightSettingsPanel.module"
@@ -15,6 +14,8 @@ import { ColorSettingsPanelModule } from "./colorSettingsPanel/colorSettingsPane
 import { CommonModule } from "@angular/common"
 import { MaterialModule } from "../../../material/material.module"
 import { LinkColorMetricToHeightMetricButtonModule } from "./linkColorMetricToHeightMetricButton/linkColorMetricToHeightMetricButton.module"
+import { SearchPanelModule } from "./searchPanel/searchPanel.module"
+import { RibbonBarPanelModule } from "./ribbonBarPanel/ribbonBarPanel.module"
 
 @NgModule({
     imports: [
@@ -32,7 +33,8 @@ import { LinkColorMetricToHeightMetricButtonModule } from "./linkColorMetricToHe
         ColorMetricChooserModule,
         ColorSettingsPanelModule,
         EdgeMetricChooserModule,
-        EdgeSettingsPanelModule
+        EdgeSettingsPanelModule,
+        RibbonBarPanelModule
     ],
     declarations: [RibbonBarComponent],
     exports: [RibbonBarComponent]