+++ visualization/app/codeCharta/ui/toolBar/globalConfigurationButton/globalConfigurationButton.component.ts
@@ -1,12 +1,11 @@
-import { Component, ViewEncapsulation } from "@angular/core"
+import { Component } from "@angular/core"
 import { MatDialog } from "@angular/material/dialog"
 import { GlobalConfigurationDialogComponent } from "./globalConfigurationDialog/globalConfigurationDialog.component"
 
 @Component({
     selector: "cc-global-configuration-button",
     templateUrl: "./globalConfigurationButton.component.html",
-    styleUrls: ["./globalConfigurationButton.component.scss"],
-    encapsulation: ViewEncapsulation.None
+    styleUrls: ["./globalConfigurationButton.component.scss"]
 })
 export class GlobalConfigurationButtonComponent {
     constructor(private dialog: MatDialog) {}